fix(test page): force static rendering
