+++ apps/store/src/app/[locale]/test/page.tsx
@@ -35,3 +35,5 @@ const Page = async (props: Props) => {
   )
 }
 export default Page
+
+export const dynamic = 'force-static'
